Separate operators from functions
