+++ cmpfunc_cpp/include/meson.build
@@ -1,6 +1,6 @@
 # Include header files
 
 install_headers(
-  ['./complex_functions.h'],
+  ['./complex_operators.h', './complex_functions.h'],
   subdir: 'cmpx_spfunc'
 )
